+++ src/Jobs/CreateSearchDocument.php
@@ -28,6 +28,9 @@ class CreateSearchDocument implements ShouldQueue
     {
         //assign
         $this->searchable = $searchable;
+
+        //select queue
+        $this->onQueue(config('search.tubes.created', 'default'));
     }
 
     /**